requirements.txt
# ── Web framework ─────────────────────────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.6
sse-starlette>=2.1.0

# ── Data validation ───────────────────────────────────────────────────────────
pydantic==2.9.2
pydantic-settings==2.5.2

# ── Database (SQLite async) ───────────────────────────────────────────────────
sqlalchemy[asyncio]==2.0.35
aiosqlite>=0.19.0

# ── Auth ──────────────────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart>=0.0.9

# ── Claude API ────────────────────────────────────────────────────────────────
anthropic>=0.34.0

# ── Embeddings ────────────────────────────────────────────────────────────────
sentence-transformers>=3.1.0
numpy>=1.26.0,<2.0.0

# ── Text splitting (only the splitter — no full langchain needed) ─────────────
langchain-text-splitters>=0.3.0

# ── Excel parsing ─────────────────────────────────────────────────────────────
pandas>=2.2.0
openpyxl>=3.1.0
xlrd>=2.0.1

# ── PDF parsing ───────────────────────────────────────────────────────────────
pypdf>=4.3.0
pdfplumber>=0.11.0

# ── Storage ───────────────────────────────────────────────────────────────────
boto3>=1.35.0
supabase>=2.8.0

# ── Cache ─────────────────────────────────────────────────────────────────────
redis[hiredis]>=5.1.0

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv>=1.0.0
aiofiles>=24.1.0
httpx>=0.27.0
